services/ai-engine/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
qdrant-client
google-generativeai
openai
anthropic
grpcio
grpcio-tools
python-multipart
pypdf
python-docx
redis
asyncpg
